pyproject.toml
[tool.poetry]
name = "Amedice"
version = "0.1.0"
description = "Aggregator of medical institutions"
authors = ["Amedice Corp <amedice.main@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
django-cors-headers = "4.3.0"
django-debug-toolbar = "4.2.0"
djangorestframework = "3.14.0"
djangorestframework-jwt = "1.11.0"
djangorestframework-simplejwt = "5.3.1"
dj_database_url = "2.1.0"
drf-spectacular = "0.26.5"
drf-yasg = "1.21.7"
gunicorn = "22.0.0"
Pillow = "10.3.0"
psycopg2-binary = "2.9.9"
pydantic-settings = "2.2.1"
pydantic-settings-yaml = "0.2.0"
transliterate = "1.10.2"
isort = "^5.13.2"
black = "^24.10.0"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core>=1.5.1"]
build-backend = "poetry.core.masonry.api"
